+++ osrelease_test.go
@@ -18,7 +18,7 @@ func TestVersion(t *testing.T) {
 	require.NoError(t, err)
 	require.Greater(t, v.Major(), 0)
 
-	require.Equal(t, out, v.Original())
+	require.Equal(t, out, v.String())
 	require.Equal(
 		t, out, fmt.Sprintf(
 			"%d.%d.%d%s",
@@ -36,7 +36,7 @@ func TestNewVersion(t *testing.T) {
 	require.Equal(t, 2, v.Minor())
 	require.Equal(t, 3, v.Patch())
 	require.Empty(t, v.Suffix())
-	require.Equal(t, "1.2.3", v.Original())
+	require.Equal(t, "1.2.3", v.String())
 }
 
 func TestNewVersionWithSuffix(t *testing.T) {
@@ -45,7 +45,7 @@ func TestNewVersionWithSuffix(t *testing.T) {
 	require.Equal(t, 2, v.Minor())
 	require.Equal(t, 3, v.Patch())
 	require.Equal(t, "-gabc+3", v.Suffix())
-	require.Equal(t, "1.2.3-gabc+3", v.Original())
+	require.Equal(t, "1.2.3-gabc+3", v.String())
 }
 
 func TestExclusiveCompare(t *testing.T) {